+++ Javascript/app/views/chat/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Slaq - JS</title>
-    <link href="style.css" rel="stylesheet">
+    <link href="../style.css" rel="stylesheet">
 </head>
 <body>
 <div>
@@ -20,7 +20,7 @@
     </form>
 
     <script defer>
-        require("./scripts/chat")
+        require("./chat")
     </script>
 </div>
 </body>
