+++ Javascript/app/views/login/index.html
@@ -4,8 +4,8 @@
         <meta charset="UTF-8">
         <title>Slaq - JS</title>
         <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';"/>
-        <script type="application/javascript" src="scripts/login.js" defer></script>
-        <link href="style.css" rel="stylesheet">
+        <script type="application/javascript" src="login.js" defer></script>
+        <link href="../style.css" rel="stylesheet">
     </head>
     <body>
         <form id="login">
